proxy/proxyports: move test-only code
Currently MockIPTablesManager and proxyPortsForTest are defined in
testutils.go which will also be compiled for non-test builds (due to the
lack of a _test.go file suffix), although these are only used in test
code. Move them to proxyports_test.go along with the other test code.
